fix NUMBER(12,6) for double

+++ AMFICOM/v1/database/schema/mapletty/object/mark.sql
@@ -7,10 +7,10 @@ CREATE TABLE Mark (
 --
  name VARCHAR2(128),
  description VARCHAR2(256),
- longitude NUMBER(12,6),
- latiude NUMBER(12,6),
+ longitude DOUBLE,
+ latiude DOUBLE,
  physical_link_id VARCHAR2(32) NOT NULL,
- distance NUMBER(12,6),
+ distance DOUBLE,
  city VARCHAR2(128),
  street VARCHAR2(128),
  building VARCHAR2(128),
